+++ lib/live_view_native/utils.ex
@@ -14,6 +14,9 @@ defmodule LiveViewNative.Utils do
   def stringify_format(format) when is_binary(format), do: format
   def stringify_format(format) when is_atom(format), do: Atom.to_string(format)
 
+  def normalize_layouts(nil),
+    do: []
+
   def normalize_layouts(layouts) do
     Enum.map(layouts, fn
       {format, {mod, template}} -> {format, {mod, template}}