+++ test/support/clients/switch.ex
@@ -3,5 +3,6 @@ defmodule LiveViewNativeTest.Switch do
     format: :switch,
     component: LiveViewNativeTest.Switch.Component,
     module_suffix: :Switch,
-    template_engine: LiveViewNative.Engine
+    template_engine: LiveViewNative.Engine,
+    client: LiveViewNative.Switch.Client
 end