+++ test/support/clients/switch/client.ex
@@ -1,0 +1,8 @@
+defmodule LiveViewNativeTest.Switch.Client do
+  @moduledoc false
+
+  defstruct tags: %{
+    form: "LiveForm",
+    button: "Button"
+  }
+end